#1986: split mirror/mod.rs into mod + fast_path + resolver (pure motion)
De-monolith the mirror module by moving whole functions out of
mod.rs into two sibling files, with no logic, signature, or behavior
changes. mod.rs is now the supervisor: it keeps the MirrorCloneResult
enum, the MIRROR_TX_FRAME_RESERVE / MIRROR_PENDING_LIMIT constants, the
config-resolution + sampling-decision entry points (select_mirror_config,
resolve_mirror_config, mirror_sample_allows), declares the submodules,
re-exports the crate::afxdp-visible API, and hosts the colocated tests.

  fast_path.rs (HOT): the descriptor-copy enqueue path —
    enqueue_mirror_clone, enqueue_sampled_mirror_clone,
    enqueue_mirror_clone_to_binding (the inline UMEM frame copy +
    TX staging), enqueue_mirror_clone_to_live,
    enqueue_admitted_mirror_clone_to_live,
    enqueue_sampled_mirror_clone_to_live.

  resolver.rs (COLD): target-binding lookup and accounting —
    mirror_target_binding_index, admit_mirror_clone_to_live,
    mirror_cos_queue_id, record_mirror_clone_result.

There is no vlan.rs: mirror.rs contained no VLAN injection/replication
code (vlan only appears as a parameter name and as test fixture data),
and resolve_ingress_logical_ifindex lives in afxdp/forwarding/mod.rs,
not here — so the issue's suggested vlan.rs/resolver-ifindex moves had no
source to extract. The split honestly tracks the file's real content.

Inlining-neutrality (this is the hot mirror-clone path): every existing
#[inline] attribute is preserved. mirror_cos_queue_id and
record_mirror_clone_result keep theirs. #[inline] is ADDED only to the
two helpers whose previously same-module calls now cross the
fast_path -> resolver boundary: mirror_target_binding_index and
admit_mirror_clone_to_live. objdump of the release binary confirms the
hot path is unchanged: enqueue_sampled_mirror_clone still fully inlines
mirror_target_binding_index, admit_mirror_clone_to_live,
mirror_cos_queue_id, and enqueue_mirror_clone_to_binding (no new call
into any of them), and enqueue_admitted_mirror_clone_to_live stays the
single out-of-line cold-path call exactly as before the split.

Visibility: only one widening — mirror_target_binding_index goes from a
module-private fn to pub(super) so the fast path can reach it across the
new boundary; it stays mirror-module-internal (not added to the
crate::afxdp surface). The crate::afxdp-visible public API of
afxdp::mirror is byte-identical (13 names unchanged, diff empty).

Validation: cargo build --release clean (no warnings in mirror/);
cargo test --release passes (2057 passed, 0 failed single-threaded;
all 21 colocated mirror tests pass). One unrelated concurrency test,
afxdp::worker_queue::tests::concurrent_recovery_processes_each_command_exactly_once
(no mirror references), flakes only under heavy parallel host load and
passes in isolation and single-threaded — pre-existing, not from this
change.
