+++ userspace-dp/src/afxdp/mirror/mod.rs
@@ -1,5 +1,31 @@
 use super::*;
 
+mod fast_path;
+mod resolver;
+
+pub(in crate::afxdp) use fast_path::{
+    enqueue_admitted_mirror_clone_to_live, enqueue_sampled_mirror_clone,
+};
+// `enqueue_mirror_clone`, `enqueue_mirror_clone_to_live`, and
+// `enqueue_sampled_mirror_clone_to_live` are reachable from `crate::afxdp` but
+// are dead in non-test builds (the live/cross-worker fallbacks are exercised by
+// tests and conditional call sites). The functions themselves carry
+// `#[cfg_attr(not(test), allow(dead_code))]`; suppress the matching re-export
+// lint here, mirroring the monolithic mirror.rs glob visibility.
+#[cfg_attr(not(test), allow(unused_imports))]
+pub(in crate::afxdp) use fast_path::{
+    enqueue_mirror_clone, enqueue_mirror_clone_to_live, enqueue_sampled_mirror_clone_to_live,
+};
+pub(in crate::afxdp) use resolver::{
+    admit_mirror_clone_to_live, mirror_cos_queue_id, record_mirror_clone_result,
+};
+// Sibling-only helper (mirror-module-internal). A plain (private) `use` keeps
+// it out of the wider `crate::afxdp` surface while still letting the fast
+// path's `use super::*` resolve it across the split (children see parent-
+// private items). `pub(super) use` of a `pub(super)` item would trip E0364,
+// the same constraint documented in tx/mod.rs.
+use resolver::mirror_target_binding_index;
+
 pub(in crate::afxdp) const MIRROR_TX_FRAME_RESERVE: usize = TX_BATCH_SIZE;
 const MIRROR_PENDING_LIMIT: usize = TX_BATCH_SIZE;
 
@@ -54,358 +80,6 @@ pub(in crate::afxdp) fn mirror_sample_allows(rate: u32, sample_counter: &mut u64
         current & (rate - 1) == 0
     } else {
         current % rate == 0
-    }
-}
-
-#[cfg_attr(not(test), allow(dead_code))]
-pub(in crate::afxdp) fn enqueue_mirror_clone(
-    left: &mut [BindingWorker],
-    ingress_index: usize,
-    ingress_binding: &mut BindingWorker,
-    right: &mut [BindingWorker],
-    binding_lookup: &WorkerBindingLookup,
-    mirror_targets: &MirrorTargetMap,
-    forwarding: &ForwardingState,
-    config: MirrorRuntimeConfig,
-    ingress_queue_id: u32,
-    frame: &[u8],
-    meta: ForwardPacketMeta,
-    flow_key: Option<&SessionKey>,
-) -> MirrorCloneResult {
-    let mirror_tx_ifindex = resolve_tx_binding_ifindex(forwarding, config.output_ifindex);
-    let target_binding_index = mirror_target_binding_index(
-        binding_lookup,
-        ingress_index,
-        ingress_binding.ifindex,
-        ingress_queue_id,
-        mirror_tx_ifindex,
-    );
-    let cos_queue_id = mirror_cos_queue_id(forwarding, config.output_ifindex, meta, flow_key);
-    let Some(target_binding) = target_binding_index
-        .and_then(|idx| binding_by_index_mut(left, ingress_index, ingress_binding, right, idx))
-    else {
-        return enqueue_mirror_clone_to_live(
-            mirror_targets,
-            config,
-            mirror_tx_ifindex,
-            ingress_queue_id,
-            frame,
-            meta,
-            flow_key,
-            cos_queue_id,
-        );
-    };
-
-    enqueue_mirror_clone_to_binding(target_binding, config, frame, meta, flow_key, cos_queue_id)
-}
-
-pub(in crate::afxdp) fn enqueue_sampled_mirror_clone(
-    left: &mut [BindingWorker],
-    ingress_index: usize,
-    ingress_binding: &mut BindingWorker,
-    right: &mut [BindingWorker],
-    binding_lookup: &WorkerBindingLookup,
-    mirror_targets: &MirrorTargetMap,
-    forwarding: &ForwardingState,
-    ingress_ifindex: i32,
-    ingress_vlan_id: u16,
-    ingress_queue_id: u32,
-    frame: &[u8],
-    meta: ForwardPacketMeta,
-    flow_key: Option<&SessionKey>,
-) -> Option<MirrorCloneResult> {
-    let config = resolve_mirror_config(forwarding, ingress_ifindex, ingress_vlan_id)?;
-    let mirror_tx_ifindex = resolve_tx_binding_ifindex(forwarding, config.output_ifindex);
-    let target_binding_index = mirror_target_binding_index(
-        binding_lookup,
-        ingress_index,
-        ingress_binding.ifindex,
-        ingress_queue_id,
-        mirror_tx_ifindex,
-    );
-    let cos_queue_id = mirror_cos_queue_id(forwarding, config.output_ifindex, meta, flow_key);
-    if let Some(target_binding_index) = target_binding_index {
-        if !mirror_sample_allows(config.rate, &mut ingress_binding.mirror_sample_counter) {
-            return None;
-        }
-        let Some(target_binding) = binding_by_index_mut(
-            left,
-            ingress_index,
-            ingress_binding,
-            right,
-            target_binding_index,
-        ) else {
-            return Some(MirrorCloneResult::NoBinding);
-        };
-        return Some(enqueue_mirror_clone_to_binding(
-            target_binding,
-            config,
-            frame,
-            meta,
-            flow_key,
-            cos_queue_id,
-        ));
-    } else {
-        let admission = match admit_mirror_clone_to_live(
-            mirror_targets,
-            mirror_tx_ifindex,
-            ingress_queue_id,
-            frame.len(),
-        ) {
-            Ok(admission) => admission,
-            Err(result) => return Some(result),
-        };
-        if !mirror_sample_allows(config.rate, &mut ingress_binding.mirror_sample_counter) {
-            return None;
-        }
-        return Some(enqueue_admitted_mirror_clone_to_live(
-            admission,
-            config,
-            frame.to_vec(),
-            meta,
-            flow_key,
-            cos_queue_id,
-        ));
-    }
-}
-
-fn enqueue_mirror_clone_to_binding(
-    target_binding: &mut BindingWorker,
-    config: MirrorRuntimeConfig,
-    frame: &[u8],
-    meta: ForwardPacketMeta,
-    flow_key: Option<&SessionKey>,
-    cos_queue_id: Option<u8>,
-) -> MirrorCloneResult {
-    if frame.len() > tx_frame_capacity() {
-        return MirrorCloneResult::NoFrame;
-    }
-    let pending_mirror_pressure = target_binding
-        .tx_pipeline
-        .pending_tx_prepared
-        .len()
-        .saturating_add(target_binding.tx_pipeline.pending_tx_local.len());
-    if pending_mirror_pressure >= MIRROR_PENDING_LIMIT {
-        return MirrorCloneResult::QueueFullSameWorker;
-    }
-    if target_binding.tx_pipeline.free_tx_frames.len() <= MIRROR_TX_FRAME_RESERVE {
-        return MirrorCloneResult::TxFrameReserve;
-    }
-    let Some(tx_offset) = target_binding.tx_pipeline.free_tx_frames.pop_front() else {
-        return MirrorCloneResult::TxFrameReserve;
-    };
-    let Some(out) = (unsafe {
-        target_binding
-            .umem
-            .area()
-            .slice_mut_unchecked(tx_offset as usize, frame.len())
-    }) else {
-        target_binding
-            .tx_pipeline
-            .free_tx_frames
-            .push_front(tx_offset);
-        return MirrorCloneResult::NoFrame;
-    };
-    out.copy_from_slice(frame);
-    target_binding
-        .tx_pipeline
-        .pending_tx_prepared
-        .push_back(PreparedTxRequest {
-            offset: tx_offset,
-            len: frame.len() as u32,
-            recycle: PreparedTxRecycle::FreeTxFrame,
-            expected_ports: None,
-            expected_addr_family: meta.addr_family,
-            expected_protocol: meta.protocol,
-            flow_key: flow_key.cloned(),
-            egress_ifindex: config.output_ifindex,
-            cos_queue_id,
-            dscp_rewrite: None,
-            mirror_clone: true,
-            enqueue_ns: 0,
-        });
-    MirrorCloneResult::Enqueued
-}
-
-fn mirror_target_binding_index(
-    binding_lookup: &WorkerBindingLookup,
-    ingress_index: usize,
-    ingress_ifindex: i32,
-    ingress_queue_id: u32,
-    mirror_tx_ifindex: i32,
-) -> Option<usize> {
-    if ingress_ifindex == mirror_tx_ifindex {
-        return Some(ingress_index);
-    }
-    binding_lookup
-        .by_if_queue
-        .get(&(mirror_tx_ifindex, ingress_queue_id))
-        .copied()
-        .or_else(|| {
-            let indices = binding_lookup.all_by_if.get(&mirror_tx_ifindex)?;
-            (indices.len() == 1).then_some(indices[0])
-        })
-}
-
-#[cfg_attr(not(test), allow(dead_code))]
-pub(in crate::afxdp) fn enqueue_mirror_clone_to_live(
-    mirror_targets: &MirrorTargetMap,
-    config: MirrorRuntimeConfig,
-    mirror_tx_ifindex: i32,
-    ingress_queue_id: u32,
-    frame: &[u8],
-    meta: ForwardPacketMeta,
-    flow_key: Option<&SessionKey>,
-    cos_queue_id: Option<u8>,
-) -> MirrorCloneResult {
-    let admission = match admit_mirror_clone_to_live(
-        mirror_targets,
-        mirror_tx_ifindex,
-        ingress_queue_id,
-        frame.len(),
-    ) {
-        Ok(admission) => admission,
-        Err(result) => return result,
-    };
-    enqueue_admitted_mirror_clone_to_live(
-        admission,
-        config,
-        frame.to_vec(),
-        meta,
-        flow_key,
-        cos_queue_id,
-    )
-}
-
-pub(in crate::afxdp) fn admit_mirror_clone_to_live(
-    mirror_targets: &MirrorTargetMap,
-    mirror_tx_ifindex: i32,
-    ingress_queue_id: u32,
-    frame_len: usize,
-) -> Result<PendingTxAdmission, MirrorCloneResult> {
-    if frame_len > tx_frame_capacity() {
-        return Err(MirrorCloneResult::NoFrame);
-    }
-    let Some(target_live) = mirror_targets.target_live(mirror_tx_ifindex, ingress_queue_id) else {
-        return Err(MirrorCloneResult::NoBinding);
-    };
-    BindingLiveState::try_reserve_mirror_tx_owned(&target_live, MIRROR_PENDING_LIMIT)
-        .map_err(|_| MirrorCloneResult::QueueFullCrossWorker)
-}
-
-pub(in crate::afxdp) fn enqueue_admitted_mirror_clone_to_live(
-    admission: PendingTxAdmission,
-    config: MirrorRuntimeConfig,
-    frame: Vec<u8>,
-    meta: ForwardPacketMeta,
-    flow_key: Option<&SessionKey>,
-    cos_queue_id: Option<u8>,
-) -> MirrorCloneResult {
-    if frame.len() > tx_frame_capacity() {
-        return MirrorCloneResult::NoFrame;
-    }
-    let req = TxRequest {
-        bytes: frame,
-        expected_ports: None,
-        expected_addr_family: meta.addr_family,
-        expected_protocol: meta.protocol,
-        flow_key: flow_key.cloned(),
-        egress_ifindex: config.output_ifindex,
-        cos_queue_id,
-        dscp_rewrite: None,
-        mirror_clone: true,
-        enqueue_ns: 0,
-    };
-    admission
-        .enqueue_owned(req)
-        .map(|_| MirrorCloneResult::Enqueued)
-        .unwrap_or(MirrorCloneResult::QueueFullCrossWorker)
-}
-
-#[cfg_attr(not(test), allow(dead_code))]
-pub(in crate::afxdp) fn enqueue_sampled_mirror_clone_to_live(
-    live: &BindingLiveState,
-    mirror_targets: &MirrorTargetMap,
-    forwarding: &ForwardingState,
-    ingress_ifindex: i32,
-    ingress_vlan_id: u16,
-    ingress_queue_id: u32,
-    sample_counter: &mut u64,
-    frame: &[u8],
-    meta: ForwardPacketMeta,
-    flow_key: Option<&SessionKey>,
-) -> Option<MirrorCloneResult> {
-    let config = resolve_mirror_config(forwarding, ingress_ifindex, ingress_vlan_id)?;
-    let cos_queue_id = mirror_cos_queue_id(forwarding, config.output_ifindex, meta, flow_key);
-    let admission = match admit_mirror_clone_to_live(
-        mirror_targets,
-        resolve_tx_binding_ifindex(forwarding, config.output_ifindex),
-        ingress_queue_id,
-        frame.len(),
-    ) {
-        Ok(admission) => admission,
-        Err(result) => {
-            record_mirror_clone_result(live, result, frame.len());
-            return Some(result);
-        }
-    };
-    if !mirror_sample_allows(config.rate, sample_counter) {
-        return None;
-    }
-    let result = enqueue_admitted_mirror_clone_to_live(
-        admission,
-        config,
-        frame.to_vec(),
-        meta,
-        flow_key,
-        cos_queue_id,
-    );
-    record_mirror_clone_result(live, result, frame.len());
-    Some(result)
-}
-
-#[inline]
-pub(in crate::afxdp) fn mirror_cos_queue_id(
-    forwarding: &ForwardingState,
-    output_ifindex: i32,
-    meta: ForwardPacketMeta,
-    flow_key: Option<&SessionKey>,
-) -> Option<u8> {
-    resolve_cached_cos_tx_selection(forwarding, output_ifindex, meta.into(), flow_key).queue_id
-}
-
-#[inline]
-pub(in crate::afxdp) fn record_mirror_clone_result(
-    live: &BindingLiveState,
-    result: MirrorCloneResult,
-    frame_len: usize,
-) {
-    match result {
-        MirrorCloneResult::Enqueued => {
-            live.mirrored_packets.fetch_add(1, Ordering::Relaxed);
-            live.mirrored_bytes
-                .fetch_add(frame_len as u64, Ordering::Relaxed);
-        }
-        MirrorCloneResult::NoBinding => {
-            live.mirror_drops_no_binding.fetch_add(1, Ordering::Relaxed);
-        }
-        MirrorCloneResult::NoFrame => {
-            live.mirror_drops_no_frame.fetch_add(1, Ordering::Relaxed);
-        }
-        MirrorCloneResult::TxFrameReserve => {
-            live.mirror_drops_tx_frame_reserve
-                .fetch_add(1, Ordering::Relaxed);
-        }
-        MirrorCloneResult::QueueFullSameWorker => {
-            live.mirror_drops_queue_full.fetch_add(1, Ordering::Relaxed);
-            live.mirror_drops_queue_full_same_worker
-                .fetch_add(1, Ordering::Relaxed);
-        }
-        MirrorCloneResult::QueueFullCrossWorker => {
-            live.mirror_drops_queue_full.fetch_add(1, Ordering::Relaxed);
-            live.mirror_drops_queue_full_cross_worker
-                .fetch_add(1, Ordering::Relaxed);
-        }
     }
 }
